backend/pyproject.toml
[project]
name = "backend-das"
version = "0.1.0"
description = "Django Rest Server for DAS"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-stubs>=5.2.8",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "ollama>=0.6.1",
    "pylint-django>=2.7.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.0",
]
